workers/python/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "musicforge-worker"
version = "0.1.0"
description = "AI-powered music generation worker for MusicForge"
readme = "README.md"
requires-python = ">=3.9"
license = {text = "MIT"}
authors = [
    {name = "MusicForge Team"}
]

dependencies = [
    # gRPC
    "grpcio>=1.60.0",
    "grpcio-tools>=1.60.0",
    "protobuf>=4.25.0",
    
    # AI Models
    "torch>=2.0.0",
    "torchaudio>=2.0.0",
    "transformers>=4.36.0",
    # audiocraft installed manually in Dockerfile to bypass torchtext conflict
    "bark>=0.1.5",        # Voice synthesis
    "demucs>=4.0.0",      # Stem separation
    
    # Audio processing
    "scipy>=1.11.0",
    "numpy>=1.26.0",
    "soundfile>=0.12.0",
    "librosa>=0.10.0",
    
    # Utilities
    "pydantic>=2.5.0",
    "python-dotenv>=1.0.0",
    "structlog>=24.1.0",
    "music21>=8.1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.1.0",
    "ruff>=0.1.0",
    "mypy>=1.7.0",
]

[project.scripts]
musicforge-worker = "src.server:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.11"
strict = true
ignore_missing_imports = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
